UNLP image sin dependencias instaladas/requirements.txt
matplotlib-inline==0.1.6
notebook==6.5.4
pysimplegui-exemaker==1.3.0
streamlit==1.23.1
wordcloud==1.9.2